requirements.txt
adblockparser==0.7
aliyun-python-sdk-core-v3==2.13.0
aliyun-python-sdk-kms==2.5.0
asn1crypto==0.24.0
astroid==2.1.0
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.3
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
constantly==15.1.0
crcmod==1.7
cryptography==2.5
cssselect==1.0.3
dateparser==0.7.1
funcparserlib==0.3.6
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
lxml==4.3.1
mccabe==0.6.1
oss2==2.6.1
parsel==1.5.1
Pillow==5.4.1
psutil==5.5.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.7.3
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.2.2
pymongo==3.7.2
pyOpenSSL==19.0.0
pyquery==1.4.0
python-dateutil==2.8.0
pytz==2018.9
qt5reactor==0.5
queuelib==1.5.0
redis==3.2.1
regex==2019.2.21
requests==2.21.0
Scrapy==1.6.0
scrapy-splash==0.7.2
scrapyd==1.2.0
scrapyd-client==1.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
splash==3.3.1
SQLAlchemy==1.2.18
stem==1.7.1
Twisted==18.9.0
typed-ast==1.3.1
tzlocal==1.5.1
urllib3==1.24.1
w3lib==1.20.0
wrapt==1.11.1
xvfbwrapper==0.2.9
zope.interface==4.6.0
